DB015: Expand domain Models for Sessions
