+++ core/domain/src/commonMain/kotlin/pt/dourobats/app/core/domain/model/SessionStatus.kt
@@ -1,0 +1,10 @@
+package pt.dourobats.app.core.domain.model
+
+/**
+ * Represents the session state.
+ */
+enum class SessionStatus {
+    SCHEDULED,
+    CANCELLED,
+    COMPLETED
+}
