+++ core/domain/src/commonMain/kotlin/pt/dourobats/app/core/domain/model/SkillLevel.kt
@@ -1,0 +1,10 @@
+package pt.dourobats.app.core.domain.model
+
+/**
+ * Represents the skill level targeted by a session.
+ */
+enum class SkillLevel {
+    BEGINNER,
+    INTERMEDIATE,
+    ADVANCED
+}
