+++ core/domain/src/commonMain/kotlin/pt/dourobats/app/core/domain/model/Sport.kt
@@ -1,0 +1,13 @@
+package pt.dourobats.app.core.domain.model
+
+/**
+ * Represents a sport that can be associated with training sessions.
+ */
+data class Sport (
+
+    val id: String,
+    val name: String,
+    val description: String,
+    val iconResource: String
+
+)
