+++ core/domain/src/commonMain/kotlin/pt/dourobats/app/core/domain/repository/TrainingRepository.kt
@@ -1,12 +1,12 @@
 package pt.dourobats.app.core.domain.repository
 
 import kotlinx.coroutines.flow.Flow
-import pt.dourobats.app.core.domain.model.TrainingSession
+
 
 /**
  * Repository interface for training session operations
  * Following Repository Pattern from Clean Architecture
- */
+
 interface TrainingRepository {
     /**
      * Get all training sessions as a Flow
@@ -23,3 +23,4 @@ interface TrainingRepository {
      */
     suspend fun saveTrainingSession(session: TrainingSession)
 }
+ */
